+++ src/de/dhbw/Microcontroller/Befehle/PIC/IORLW.java
@@ -15,12 +15,12 @@ public class IORLW extends Instruction {
 
     @Override
     public void execute(){
-        byte k = (byte) argument;
+        int k = argument;
 
-        memory.setRegisterW((byte) (memory.getRegisterW() | k));
+        memory.setRegisterW(memory.getRegisterW() | k);
 
         // CheckZero
-        Byte w = memory.getRegisterW();
+        int w = memory.getRegisterW();
         if(w == 0)
             setZeroFlag();
         else
